+++ packages/suparepo/lib/src/edge_function_generator.dart
@@ -12,8 +12,6 @@ class EdgeFunctionGenerator {
     List<EdgeFunctionInfo> functions, {
     Map<String, EdgeFunctionModelDef>? modelDefs,
     String supabaseImport = defaultSupabaseImport,
-    bool generateProviders = false,
-    String? clientProviderImport,
   }) {
     final buffer = StringBuffer();
 
@@ -23,11 +21,6 @@ class EdgeFunctionGenerator {
     buffer.writeln();
 
     // Imports
-    if (generateProviders) {
-      buffer.writeln(
-        "import 'package:riverpod_annotation/riverpod_annotation.dart';",
-      );
-    }
     buffer.writeln("import '$supabaseImport';");
 
     // Check if any function has model definitions
@@ -37,15 +30,6 @@ class EdgeFunctionGenerator {
       buffer.writeln(
         "import 'dart:convert';",
       );
-    }
-
-    if (generateProviders) {
-      buffer.writeln();
-      final providerImport =
-          clientProviderImport ?? 'supabase_client_provider.dart';
-      buffer.writeln("import '$providerImport';");
-      buffer.writeln();
-      buffer.writeln("part 'edge_function_client.g.dart';");
     }
 
     buffer.writeln();
@@ -100,27 +84,6 @@ class EdgeFunctionGenerator {
     }
 
     buffer.writeln('}');
-
-    // Generate Riverpod provider if enabled
-    if (generateProviders) {
-      buffer.writeln();
-      buffer.writeln(
-        '/// Provider for [SupabaseEdgeFunctionClient]',
-      );
-      buffer.writeln('@Riverpod(keepAlive: true)');
-      buffer.writeln(
-        'SupabaseEdgeFunctionClient '
-        'supabaseEdgeFunctionClient(Ref ref) {',
-      );
-      buffer.writeln(
-        '  final client = '
-        'ref.watch(supabaseClientProvider);',
-      );
-      buffer.writeln(
-        '  return SupabaseEdgeFunctionClient(client);',
-      );
-      buffer.writeln('}');
-    }
 
     return buffer.toString();
   }